+++ project1/static/styles/details.css
@@ -25,7 +25,7 @@ body{
 	text-align : center;
 	font-family : Georgia, 'Times New Roman', Times, serif;
 	font-size : 40px;
-	color : aqua;
+	color : #ffe600;
 }
 
 .list{
@@ -60,9 +60,9 @@ body{
 .Image{
 	padding : 10px;
 	margin : 10px;
-	height : 70%;
+	height : 50%;
 	background-color : transparent;
-	width : 25%;
+	width : 20%;
 	border : 2px;
 }
